requirements.txt
click==8.1.7
joblib==1.4.2
nltk==3.9.1
pillow==10.4.0
regex==2024.9.11
tqdm==4.66.5
ttkbootstrap==1.10.1
ufal.morphodita==1.11.2.1
